requirements.txt
gradio
opencv-python
pillow
tqdm
python-dotenv
google-search-results
